+++ backend/src/main/java/apirest/gympass/entityDto/EventoDTO.java
@@ -1,6 +1,5 @@
 package apirest.gympass.entityDto;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.AllArgsConstructor;
@@ -11,17 +10,17 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class EventoDTO {
-    private Long        idEvento;
-    private String      nombre;
-    private String      descripcion;
-    private LocalDate   fechaInicio;
-    private Integer     duracion;
-    private String      direccion;
-    private String      estado;
-    private String      destacado;
-    private Integer     aforoMaximo;
-    private Integer     minimoAsistencia;
-    private BigDecimal  precio;
-    private Long        idTipo;
-    private String      nombreTipo;
+    private Long idEvento;
+    private String nombre;
+    private String descripcion;
+    private LocalDate fechaInicio;
+    private Integer duracion;
+    private String direccion;
+    private String estado;
+    private String destacado;
+    private Integer aforoMaximo;
+    private Integer minimoAsistencia;
+    private float precio;
+    private Long idTipo;
+    private String nombreTipo;
 }
